% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_convert_files.R
\name{convert_files}
\alias{convert_files}
\title{Convert files to a specified format}
\usage{
convert_files(
  folder,
  files,
  new_extension = "parquet",
  new_folder = NULL,
  verbose = TRUE
)
}
\arguments{
\item{folder}{The path to the folder containing the files to be converted.}

\item{files}{A character vector of the file names to be converted. The valid
file extensions are the ones supported by `dataRC::read_fun`.}

\item{new_extension}{The extension of the new format to which the files will
be converted. The valid extensions are the ones supported by
`dataRC::write_fun`.}

\item{new_folder}{The path to the folder where the converted files will be
saved. If not provided, the new folder will be created inside the original
folder with the same name.}

\item{verbose}{Logical (default is TRUE) indicating whether to display
progress messages.}
}
\description{
This function converts files to a new format. It iterates through the
provided list of files and converts each file to the specified format.
}
\details{
This function is robust to uppercase and dots in the new extension.
If the `new_folder` argument is not provided, the new folder will be created
inside the original folder with the same name. Subfolders will be created
inside the new folder to mirror the structure of the original folder.
}
\examples{
\dontrun{
# Convert files in the current directory to Parquet format
convert_files(folder = ".", files = list.files(), new_extension = "parquet")
}
}
