% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{crosstab}
\alias{crosstab}
\title{Cross Tabulation}
\usage{
crosstab(df, var1, var2)
}
\arguments{
\item{df}{A dataframe.}

\item{var1}{A character with the name of the first variable .}

\item{var2}{A character with the name  of the second variable.}
}
\value{
A cross-tabulation table with the following properties:
\itemize{
\item The first column has the name of \code{var1} and all values of \code{var1} in \code{df}.
\item The other columns has as names all the values of \code{var2} in \code{df}.
\item The cell (i, j) has the absolute frequency of the combination of values.
}
}
\description{
Performs a cross-tabulation of two variables in a dataframe.
}
\examples{
crosstab(mtcars, "gear", "carb")
}
