% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_delete_files.R
\name{delete_files_from_excel}
\alias{delete_files_from_excel}
\title{Delete files based on an Excel file}
\usage{
delete_files_from_excel(file_path, save = 0, verbose = T)
}
\arguments{
\item{file_path}{The path to the Excel file containing file details.}

\item{save}{The value in the delete column that avoid the deletion of the
file. The default value is 0.}

\item{verbose}{If TRUE (default), displays a message to the user.}
}
\value{
None
}
\description{
Based on an excel file, this function deletes all the files oriented by the
user.
}
\details{
This function should be used with `dataR::folder_details()`, which generates
an Excel file with detailed information for each file of a folder. After
this, the user must manually modify the "delete" column in such way, that
the files that are expected to be deleted must have any value different to the
`save` value (by default 0).
}
\examples{
\dontrun{
delete_files_from_excel("details.xlsx")
}
}
