% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_delete_files.R
\name{delete_stata_temps}
\alias{delete_stata_temps}
\title{Delete Stata temporary files}
\usage{
delete_stata_temps()
}
\value{
None.
}
\description{
The function identifies Stata temporary files based on their naming
convention. It then deletes the identified files and prints the number of
files deleted along with the cleared disk space.
}
\examples{
\dontrun{
delete_stata_temps()
}
}
\seealso{
\href{https://www.stata.com/statalist/archive/2004-01/msg00542.html}{Stata temporary files FAQ}
}
