% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_size_of_files.R
\name{files_size}
\alias{files_size}
\title{Get the total size of multiple files.}
\usage{
files_size(files, units = "bytes")
}
\arguments{
\item{files}{A vector of file paths.}

\item{units}{The desired file size unit: \code{'bytes'} (the default), \code{'kb'},
\code{'mb'}, or \code{'gb'}.}
}
\value{
The total size of all files combined.
}
\description{
This function takes a vector of file paths and returns the total size of all
files combined. File sizes can be reported in different units (bytes,
kilobytes, megabytes, gigabytes).
}
\examples{
\dontrun{
files_size(c("file1.txt", "file2.txt"), units = 'mb')
}
}
