% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_delete_files.R
\name{find_stata_tempdir}
\alias{find_stata_tempdir}
\title{Find Stata Temporary Directory}
\usage{
find_stata_tempdir()
}
\value{
A string representing the path to the Stata temporary directory.
}
\description{
This function attempts to find the directory where Stata stores its temporary
files. It checks the \code{STATATMP} environment variable first, since is used to
modify predefined temporary files. If \code{STATATMP} doesn't exists, it falls
back to \code{TEMP} on Windows or \code{TMPDIR} on Linux/Unix.
}
\examples{
\dontrun{
find_stata_tempdir()
}
}
\seealso{
\href{https://www.stata.com/support/faqs/data-management/statatmp-environment-variable/#}{Using the STATATMP environment variable}
}
