% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_delete_files.R
\name{folder_details}
\alias{folder_details}
\title{Create a table with information of all the files inside a list of folders}
\usage{
folder_details(folders, file_path = NULL, units = "bytes", verbose = TRUE)
}
\arguments{
\item{folders}{A vector of folder paths.}

\item{file_path}{The path where the Excel file will be saved.}

\item{units}{The desired units for file sizes ('bytes', 'kb', 'mb', or 'gb').}

\item{verbose}{If TRUE (default), displays a message to the user.}
}
\value{
None
}
\description{
This function takes a vector of folder paths, retrieves details about files
in those folders, and writes the information to an Excel file. The details
include file names, last update timestamps, file sizes proportional sizes to
the folder where it is stored, and add a column for potential deletion.
}
\details{
Eases the cleaning of folder by creating an excel with the details
  about each file. Once created the user must change/delete the zeros in the
  delete column for the files that must be deleted.
}
\examples{
\dontrun{
folder_details(c("/path/to/folder1", "/path/to/folder2"),
                file_path = "details.xlsx", units = 'kb')
}
}
