% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_partial_dictionary.R
\name{get_columns_with_same_names}
\alias{get_columns_with_same_names}
\title{Get Columns with Same Names}
\usage{
get_columns_with_same_names(input_file, sheet_name = "colname")
}
\arguments{
\item{input_file}{Path to the Excel file containing column names.}

\item{sheet_name}{Name of the sheet in the Excel file containing the column
names (default is 'colname').}
}
\value{
A list where each element represents a unique column name, and the
  corresponding value is a vector of column names that are identical to it.
}
\description{
This function reads column names from an Excel file and identifies columns
with the same names. It returns a list where each element represents a unique
column name, and the corresponding value is a vector of column names that are
identical to it.
}
\examples{
\dontrun{
# Get columns with the same names from an Excel file
get_columns_with_same_names(input_file = "column_names.xlsx", sheet_name = "colname")
}

}
