% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{get_next_consecutives}
\alias{get_next_consecutives}
\title{Get Next Consecutive Numbers}
\usage{
get_next_consecutives(prev, post)
}
\arguments{
\item{prev}{A vector of positive integers representing previous numbers.}

\item{post}{A vector of positive integers representing subsequent numbers.}
}
\value{
A vector containing the next consecutive numbers in 'post' after each element in 'prev'.
}
\description{
Given two vectors of integers, this function finds the next consecutive number
in the second vector after each element in the first vector.
}
\examples{
\dontrun{
prev <- c(1, 3, 5)
post <- c(2, 4, 6, 7)
get_next_consecutives(prev, post)
Output: 2, 4, 6.
}

}
