% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_unify_data.R
\name{get_unifying_file_info}
\alias{get_unifying_file_info}
\title{Get necessary data for \code{\link[=unify_colnames]{unify_colnames()}} and
\code{\link[=unify_classes]{unify_classes()}}}
\usage{
get_unifying_file_info(dict, file, selected_columns)
}
\arguments{
\item{dict}{Data frame that represents a refined unifying dictionary
(possible created by \code{\link[=sort_partial_dictionary]{sort_partial_dictionary()}} and refined by the
user) that contains information about a group of files intended to be
process together. It must have at least three columns: \code{uniname},
\code{uniclass} and the \code{file} name.}

\item{file}{String that represents a file name that is part of the group of
files (i.e. is a column name in \code{dict}).}

\item{selected_columns}{Atomic vector that is a subset of the \code{uninames}
\code{dict}'s column. In other words, is a set of desired columns that
individually, should be in at least one file of the group of files.}
}
\value{
A list with three fields:
\itemize{
\item \code{original_colnames}: Atomic character vector that contains the names of
the file's columns that are related to the \code{selected_columns}'s \code{uninames}.
\item \code{new_colnames}: Atomic character vector that contains the \code{uniname}s
associated with the file variables \code{original_colnames}. Notice that are
arranged to be aligned with the \code{original_colnames}.
\item \code{new_classes}: Atomic character vector that contains the \code{uniclass}es
associated with the file variables \code{original_colnames}. Notice that are
arranged to be aligned with the \code{original_colnames}.
}
}
\description{
This function is part of a group of functions intended to solve a scenario
where there is equivalent data that is potentially stored heterogeneously
(e.g. different column names and datatypes).
}
\details{
This function returns auxiliary information that is employed by
\code{\link[=unify_colnames]{unify_colnames()}} to unify the column names and by
\code{\link[=unify_classes]{unify_classes()}} to unify the column data types across a group of
files.
}
\examples{
\dontrun{
unifying_file_info <- get_unifying_file_info(
    'dict.xlsx', 'example.parquet', c('YEAR', 'MONTH', 'SEX'))
}

}
\seealso{
For a full example, see the vignette
\code{process_data_with_partial_dict} in the
\href{https://jdrengifoc.github.io/dataRC/articles/process_data_with_partial_dict.html}{website}
or with the command \verb{vignette('process_data_with_partial_dict', package = dataRC')}.
}
