% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{get_values_tic_msg}
\alias{get_values_tic_msg}
\title{Measures elapsed time in a desired unit}
\usage{
get_values_tic_msg(unit = "sec")
}
\arguments{
\item{unit}{A string specifying the unit of time for the returned
value (the default is \code{'sec'}). Possible values: \code{'sec'} (seconds), \code{'min'}
(minutes) or \code{'hour'} (hours).}
}
\value{
A numeric value representing the elapsed time in the specified
\code{unit.}
}
\description{
Used with \code{\link[tictoc:tic]{tictoc::tic()}}, overrides \code{\link[tictoc:tic]{tictoc::toc()}} by returning the numeric
elapsed time since the last \code{\link[tictoc:tic]{tictoc::tic()}} call, in the specified \code{unit.}
}
\examples{
# Extract elapsed time in seconds
get_values_tic_msg(unit = 'sec')

# Extract elapsed time in minutes
get_values_tic_msg(unit = 'min')

# Extract elapsed time in hours
get_values_tic_msg(unit = 'hour')

}
