% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{is_subset_bidirectional}
\alias{is_subset_bidirectional}
\title{Check if Strings are subsets of each other.}
\usage{
is_subset_bidirectional(str1, str2)
}
\arguments{
\item{str1}{Atomic string vector.}

\item{str2}{Atomic string vector.}
}
\value{
A logical vector indicating whether each corresponding pair of
strings satisfies the subset condition.
}
\description{
This function checks if one string is a subset of another in either direction,
for each pair of strings in the input vectors.
}
\examples{
is_subset_bidirectional("hello", "hi")  # FALSE
is_subset_bidirectional(c("hello", "world"), c("hello world", "world hello"))  # TRUE TRUE
is_subset_bidirectional("hello", c("hello", "world"))  # TRUE FALSE
is_subset_bidirectional(c("hello", "world"), "hello world")  # TRUE TRUE
}
