% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{is_substring}
\alias{is_substring}
\title{Check if String is Substring of Another}
\usage{
is_substring(str1, str2)
}
\arguments{
\item{str1}{The first string.}

\item{str2}{The second string.}
}
\value{
TRUE if str1 is a substring of str2, otherwise FALSE.
}
\description{
This function checks if one string is a substring of another.
}
\examples{
is_substring("abc", "abcdef") # TRUE
is_substring("xyz", "abcdef") # FALSE

}
