% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{n_digits_int}
\alias{n_digits_int}
\title{Count the number of digits in an integer}
\usage{
n_digits_int(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Number of digits in each element of x.
}
\description{
Counts the number of digits in an integer.
}
\examples{
n_digits_int(-12345)
n_digits_int(0)
}
