% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{n_digits_natural}
\alias{n_digits_natural}
\title{Count number of digits in a natural number}
\usage{
n_digits_natural(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Number of digits in each element of x.
}
\description{
Counts the number of digits in a natural number (do not includes zero).
}
\examples{
n_digits_natural(12345)
}
