% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_partial_dictionary.R
\name{pad_with_nas}
\alias{pad_with_nas}
\title{Append NA values to a vector to reach a target length}
\usage{
pad_with_nas(vec, target_length)
}
\arguments{
\item{vec}{The input vector to be padded.}

\item{target_length}{The desired length of the padded vector.}
}
\value{
A vector padded with NA values to reach the target length.
}
\description{
This function pads a vector with NA values to achieve a specified target
length.
}
\examples{
# Pad a vector with NA values to reach a target length of 10
pad_with_nas(c(1, 2, 3), 10)

}
