% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_partition_data.R
\name{partition_data}
\alias{partition_data}
\title{Partition a parquet file into multiple parquet files of a maximum size.}
\usage{
partition_data(
  original_file,
  partition_folder,
  max_partition_size = 25,
  units = "mb"
)
}
\arguments{
\item{original_file}{Path to the original parquet file.}

\item{partition_folder}{Path to the folder where partitions will be saved.}

\item{max_partition_size}{Maximum size of each partition (default is 25).}

\item{units}{Units of the maximum size ('bytes', 'kb', 'mb', 'gb').}
}
\value{
None
}
\description{
Why do not do it for more functions? ADD VERBOSE
This function partitions a parquet file into multiple partitions based on the
specified maximum partition size.
}
\note{
In the urge enhance the performance, the size of each partition is
  forecast by assuming homogeneous storage demand along the original file.
  However this may be unrealistic, thus, the `max_partition_size` do not
  guarantee that the partition with the largest size have at most this size.
  The above is specially true for small files/partitions, since the memory
  gains due to the use of parquet becomes weaker.
}
\examples{
\dontrun{
partition_data("data/original_data.csv", "partitions/",
               max_partition_size = 25, units = 'mb')
}

}
