% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_convert_files.R
\name{read_fun}
\alias{read_fun}
\title{Wrapper function for reading data functions}
\usage{
read_fun(extension, ...)
}
\arguments{
\item{extension}{A string specifying the file extension that must support the
returned function. Supported extensions: \code{'parquet'}, \code{'feather'}, \code{'dta'},
\code{'xlsx'}, \code{'csv'}, \code{'sas'}, \code{'rds'} and \code{'rdata'}. This parameter is case
robust.}

\item{...}{Additional arguments to be passed to the reading functions
encapsulated in \code{\link[=read_fun]{read_fun()}}.}
}
\value{
A function that can read data from the specified file format. The
returned function will accept the file path as its first argument.
}
\description{
This function receives a file extension and returns the proper reading
function from popular R packages such as: \code{arrow}, \code{haven}, \code{readxl} and
\code{utils}.
}
\examples{
\dontrun{
# Read Parquet file
reader <- read_fun("parquet")
data <- reader("file.parquet")

# Read Excel file
reader <- read_fun("xlsx")
data <- reader("file.xlsx")

# Read Parquet file
reader <- read_fun("parquet")
data <- reader("file.parquet")

# Read Excel file
reader <- read_fun("xlsx")
data <- reader("file.xlsx")

# Read CSV file with comma separator
reader <- read_fun("csv", sep = ',')
data <- reader("file.csv")

# Read TXT file with tab separator and specifying additional arguments
reader <- read_fun("txt", sep = '\t', skip = 1, header = FALSE)
data <- reader("file.txt")
}
}
