% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{shared_elements}
\alias{shared_elements}
\title{Check if two atoms have the same elements.}
\usage{
shared_elements(atom1, atom2)
}
\arguments{
\item{atom1}{First atom.}

\item{atom2}{Second atom.}
}
\value{
TRUE if the atoms have the same elements, FALSE otherwise.
}
\description{
Checks if two atoms have the same elements.
}
\examples{
shared_elements(1:3, c(1, 1, 1, 3, 2, 1, 3))
}
