% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_partition_data.R
\name{unpartition_data}
\alias{unpartition_data}
\title{Join partitions into a single file}
\usage{
unpartition_data(partition_folder, new_file)
}
\arguments{
\item{partition_folder}{Path to the folder containing partition files.}

\item{new_file}{Path to the new combined file.}
}
\value{
None. Writes the merged file in \code{new_file}.
}
\description{
This function bind the partitions created by \code{\link[=partition_data]{partition_data()}} into
a single file of the desired extension type.
}
\examples{
\dontrun{
unpartition_data("partitions/", "combined_data.parquet")
}
}
