% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_convert_files.R
\name{write_fun}
\alias{write_fun}
\title{Wrapper function for writing data functions}
\usage{
write_fun(extension, ...)
}
\arguments{
\item{extension}{A string specifying the file extension that must support the
returned function. Supported extensions: \code{'parquet'}, \code{'feather'}, \code{'dta'},
\code{'xlsx'}, \code{'csv'}, \code{'sas'}, \code{'rds'} and \code{'rdata'}. This parameter is case
robust.}

\item{...}{Additional arguments to be passed to the writing functions
(write.csv, write.table, etc.).}
}
\value{
A function that can write data to the specified file format. The
returned function will accept the data object and the file path as
arguments.
}
\description{
This function receives a file extension and returns the proper writing
function.
}
\examples{
\dontrun{
# Write data to Parquet file
writer <- write_fun("parquet")
writer(data, "file.parquet")

# Write data to CSV file with comma separator
writer <- write_fun("csv", sep = ',')
writer(data, "file.csv")

# Write data to TXT file with tab separator and specifying additional arguments
writer <- write_fun("txt", sep = '\t', col.names = FALSE)
writer(data, "file.txt")
}
}
