% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_convert_files.R
\name{convert_files}
\alias{convert_files}
\title{Modify multiple files to a specified format}
\usage{
convert_files(
  folder = ".",
  files = NULL,
  new_extension = "parquet",
  new_folder = NULL,
  verbose = TRUE
)
}
\arguments{
\item{folder}{The path to the folder containing the original files.}

\item{files}{A character vector with the file paths to be copy. The paths must
satisfy the following properties:
\itemize{
\item The path must start from the \code{folder} path.
\item Have a file extensions supported by \code{\link[=read_fun]{read_fun()}}.
}}

\item{new_extension}{A string with the format that must have the copies of the
original files. Must be a extension supported by \code{\link[=write_fun]{write_fun()}}.}

\item{new_folder}{The path to the folder where the converted files will be
saved (if the folder does not exist will be created). If not provided, the
new folder will be created inside the original folder with the same name.}

\item{verbose}{Logical (the default is \code{TRUE}) indicating whether to display
progress messages.}
}
\value{
None.
}
\description{
Converts a group files located in the same folder to a new format. It iterates
through the provided vector of \code{files} and creates a copy of each file in the
specified format (or extension). The user can set the location of the new
files, which is shared across all files.
}
\details{
This function is robust to uppercase and dots in the new extension.
If the \code{new_folder} argument is not provided, the new folder will be created
inside the original folder with the same name. If needed, sub folders will
be created inside the new folder to mirror the structure of the original
folder.
}
\examples{
\dontrun{
# Convert files in the current directory to Parquet format
convert_files(folder = ".", files = list.files(), new_extension = "parquet")
}
}
