% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{crosstab}
\alias{crosstab}
\title{Cross Tabulation}
\usage{
crosstab(df, var1, var2)
}
\arguments{
\item{df}{A dataframe.}

\item{var1}{Name of the first variable as character.}

\item{var2}{Name of the second variable as character.}
}
\value{
A cross-tabulation table.
}
\description{
Performs a cross-tabulation of two variables in a dataframe.
}
\examples{
crosstab(mtcars, "gear", "carb")
}
