% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_delete_files.R
\name{folder_details}
\alias{folder_details}
\title{Create table with information of all the files in specified folders}
\usage{
folder_details(folders, file_path = NULL, units = "bytes", verbose = TRUE)
}
\arguments{
\item{folders}{A vector of folder paths.}

\item{file_path}{The path where the Excel file will be saved.}

\item{units}{The desired units for file sizes \code{'bytes'} (the default),
\code{'kb'}, \code{'mb'}, or \code{'gb'}.}

\item{verbose}{A boolean. If \code{TRUE} (the default), displays a message
confirming successful completion of the process.}
}
\value{
None.
}
\description{
This function takes a vector of folder paths, retrieves details about files
in those folders. Then writes the information to an Excel file. The details
include file names, last update timestamps, file sizes proportional sizes to
the folder where it is stored, and add a column for potential deletion.
}
\details{
Simplifies folder cleaning by generating an Excel file containing
details about each file. After creation, the user is required to modify or
delete zeros in the \code{delete} column for files to be deleted. Finally, call
\code{\link[=delete_files_from_excel]{delete_files_from_excel()}}.
}
\examples{
\dontrun{
folder_details(c("/path/to/folder1", "/path/to/folder2"),
                file_path = "details.xlsx", units = 'kb')
}
}
