% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{get_values_tic_msg}
\alias{get_values_tic_msg}
\title{Extract the elapsed time in a desired unit}
\usage{
get_values_tic_msg(unit = "sec")
}
\arguments{
\item{unit}{A character string specifying the unit of time (default is sec)
for the returned value. Possible values: 'sec' (seconds), 'min' (minutes) or
'hour' (hours).}
}
\value{
A numeric value representing the elapsed time in the specified unit.
}
\description{
This function extracts timing values from a message generated by the `tic()`
and `toc()` functions by the {tictoc} package. It calculates the elapsed time
in seconds, minutes, or hours based on the specified unit.
}
\examples{
# Extract elapsed time in seconds
get_values_tic_msg(unit = 'sec')

# Extract elapsed time in minutes
get_values_tic_msg(unit = 'min')

# Extract elapsed time in hours
get_values_tic_msg(unit = 'hour')

}
