% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{is_subset_bidirectional}
\alias{is_subset_bidirectional}
\title{Check if Strings are subsets of each other.}
\usage{
is_subset_bidirectional(str1, str2)
}
\arguments{
\item{str1, str2}{A pair of character vectors with the same \code{length}, or any
of them has a length of \code{1}. If the latter applies, then all values of the
other vector are compared with the unique value in the shorter one.}
}
\value{
A logical vector indicating whether each corresponding pair of
strings satisfies the subset condition.
}
\description{
For each pair of strings in the input vectors \code{str1} and \code{str2}, checks if
one string is a subset of another in both directions.
}
\examples{
is_subset_bidirectional("hello", "hi")  # FALSE
is_subset_bidirectional(c("hello", "world"), c("hello world", "world hello"))  # TRUE TRUE
is_subset_bidirectional("hello", c("hello", "world"))  # TRUE FALSE
is_subset_bidirectional(c("hello", "world"), "hello world")  # TRUE TRUE
}
