% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{is_substring}
\alias{is_substring}
\title{Check if a String is a Substring}
\usage{
is_substring(str1, str2)
}
\arguments{
\item{str1}{The first string.}

\item{str2}{The second string.}
}
\value{
\code{TRUE} if \code{str1} is a substring of \code{str2}, otherwise returns
\code{FALSE.}
}
\description{
This function checks if a string is a substring of another.
}
\details{
Unlike \code{\link[=grepl]{grepl()}} or \code{\link[stringr:str_detect]{stringr::str_detect()}}, is robust to invalid regular
expressions. For instance, \verb{stringr::str_detect("abcd\[", "abcd\[ef")} or
\code{stringr::str_detect("abcd]", "abcd]ef")} failed to identify that the
substrings.
}
\examples{
is_substring("abc", "abcdef") # TRUE
is_substring("xyz", "abcdef") # FALSE
is_substring("abcd]", "abcd]ef") # TRUE
is_substring("abcd[", "abcd[ef") # TRUE
}
