% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_partial_dictionary.R
\name{pad_with_nas}
\alias{pad_with_nas}
\title{Append \code{NA} values to a vector to reach a target length}
\usage{
pad_with_nas(vec, target_length)
}
\arguments{
\item{vec}{The input vector to be padded.}

\item{target_length}{The desired length of the padded vector.}
}
\value{
A vector of length \code{target_length} where its first \code{length(vec)}
elements are identical to \code{vec}. The remain tail elements are \code{NA}.
}
\description{
This function pads a vector at the tail with \code{NA} values, to achieve a
specified length target.
}
\examples{
# Pad a vector with NA values to reach a target length of 10
pad_with_nas(c(1, 2, 3), 10)

}
