% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_convert_files.R
\name{read_fun}
\alias{read_fun}
\title{Function to read data for different file extensions.}
\usage{
read_fun(extension, ...)
}
\arguments{
\item{extension}{A character string specifying the file extension of the data file.
Supported extensions: 'parquet', 'feather', 'dta', 'xlsx', 'csv', 'sas',
'rds', 'rdata'.}

\item{...}{Additional arguments to be passed to the reading functions.}
}
\value{
A function that can read data from the specified file format.
        The returned function will accept the file path as its first argument.
}
\description{
This function receives a file extension and returns the proper reading function.
}
\examples{
\dontrun{
# Read Parquet file
reader <- read_fun("parquet")
data <- reader("file.parquet")

# Read Excel file
reader <- read_fun("xlsx")
data <- reader("file.xlsx")

# Read Parquet file
reader <- read_fun("parquet")
data <- reader("file.parquet")

# Read Excel file
reader <- read_fun("xlsx")
data <- reader("file.xlsx")

# Read CSV file with comma separator
reader <- read_fun("csv", sep = ',')
data <- reader("file.csv")

# Read TXT file with tab separator and specifying additional arguments
reader <- read_fun("txt", sep = '\t', skip = 1, header = FALSE)
data <- reader("file.txt")
}
}
