% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_unify_data.R
\name{sort_partial_dictionary}
\alias{sort_partial_dictionary}
\title{Refine the "raw partial dictionary"}
\usage{
sort_partial_dictionary(old_dict_path, new_dict_path = NULL, overwrite = F)
}
\arguments{
\item{old_dict_path}{Path to the raw dictionary file.}

\item{new_dict_path}{Path to save the sorted (refined) dictionary If \code{NULL},
the original file will be \code{old_dict_path} (the default is \code{NULL}).}

\item{overwrite}{Logical indicating whether to overwrite the existing
dictionary file if \code{new_dict_path} already exists. Its default value is
\code{FALSE} to avoid undesired changes.}
}
\value{
None. The function saves the sorted dictionary file to the specified
location. The dictionary has the following columns:
\itemize{
\item \code{uniname}: Suggested unifying name for each variable. It groups across
files identical case robust variable names.
\item \code{uniclass}: Empty column that is intended to be filled manually by the
user with the unifying class for each variable. Theoretically, this could
be filled with any value, however, is recommended to use supported values
by \code{\link[=unify_classes]{unify_classes()}}, as is intended to be used with this function.
\item \code{coverage}: The percentage of files that have a match with the \code{uniname}.
\item \code{class_mode}: The class mode per \code{uniname} computed with
\code{\link[=get_mode]{get_mode()}}.
\item \code{unique_classes}: All the classes per \code{uniname}.
\item One column per file named as the file itself, excluding the folder path.
Using only the file name ensures that the dictionary remains valid even if
the dataset is moved to a different location, as long as the file structure
is preserved.
}
}
\description{
This function refine the raw partial dictionary created by
\code{\link[=create_partial_dictionary]{create_partial_dictionary()}}. It reads the raw dictionary file,
sorts the column names by frequency (across all files) and alphabetically,
and creates some columns with descriptive statistics for each variable. The
sorted (refined) dictionary is saved into a xlsx file.
}
\details{
After the creation of the dictionary, the user must manually valida
the \code{uniname} suggestions and make the necessary changes. Besides, the user
must complete the \code{uniclass} column with supported values by
\code{\link[=unify_classes]{unify_classes()}}. After this, is recommended to use the functions
\code{\link[=unify_colnames]{unify_colnames()}}, \code{\link[=unify_classes]{unify_classes()}} and
\code{\link[=relocate_columns]{relocate_columns()}} for an efficient data processing.
}
\examples{
\dontrun{
# Sort a partial dictionary file
sort_partial_dictionary(old_dict_path = "original_dictionary.xlsx",
                          overwrite = TRUE)
}

}
\seealso{
For a full example, see the vignette
\code{process_data_with_partial_dict} in the
\href{https://jdrengifoc.github.io/dataRC/articles/process_data_with_partial_dict.html}{website}
or with the command
\verb{vignette('process_data_with_partial_dict', package = dataRC')}.
}
