% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_general_purpose.R
\name{stata_month2ymd}
\alias{stata_month2ymd}
\title{Convert Stata month to date}
\usage{
stata_month2ymd(stata_month)
}
\arguments{
\item{stata_month}{Numeric vector representing Stata month format.}
}
\value{
A vector of year-months in POSIXct format.
}
\description{
Convert Stata month number to date.
}
\note{
Stata counts the months since January of 1960.
}
\examples{
stata_month2ymd(19502)
stata_month2ymd(c(19502, 19503))
}
